Rename bubble_velocity to bubble_rise_velocity
