+++ chemics/bubble_rise_velocity.py
