+++ docs/index.rst
@@ -84,7 +84,7 @@ given.
    atomic_elements
    bed_expansion_factor
    biomass_composition
-   bubble_velocity
+   bubble_rise_velocity
    chemical_equation
    choking_velocity
    conversions
